changed StudentRepository

+++ src/main/java/com/example/demo/repository/StudentRepository.java
@@ -10,7 +10,7 @@ import java.util.List;
 import java.util.Map;
 
 public class StudentRepository {
-    private Map<Long, Student> list = new HashMap<>();
+    private Map<Integer, Student> list = new HashMap<>();
 
     public List<Student> findAll(){
         return new ArrayList<>(list.values());
@@ -20,8 +20,8 @@ public class StudentRepository {
                 LocalDate.of(2001, Month.DECEMBER, 16), LocalDate.now().getYear() - 2001);
         Student s2 = new Student(2, "Lui", "lkuba19@gmail.com",
                 LocalDate.of(2008, Month.JUNE, 4), LocalDate.now().getYear() - 2008);
-        list.put(1L, s1);
-        list.put(2L, s2);
+        list.put(1, s1);
+        list.put(2, s2);
     }
 
     public Student getStudent(int id) {
